ecommerce/requirements.txt
appnope==0.1.0
decorator==4.0.9
diff-match-patch==20121119
Django==1.8.6
django-allauth==0.25.2
django-bootstrap-pagination==1.6.0
django-ckeditor==4.5.1
django-classy-tags==0.7.2
django-cms==3.2.3
django-formtools==1.0
django-import-export==0.2.8
django-localflavor==1.2
django-sekizai==0.9.0
django-suit==0.2.18
django-treebeard==4.0
djangocms-admin-style==1.1.0
docutils==0.12
gnureadline==6.3.3
html5lib==0.9999999
httplib2==0.9.2
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
oauth2==1.9.0.post1
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
ptyprocess==0.5.1
python-openid==2.2.5
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.3
stripe==1.30.0
suit==2.0.2
tablib==0.11.2
traitlets==4.2.1
